Updates, probably final fixes.
Re-edited the element default values. Somehow they got changed back.

Also made it so that when you open up the attributes editor by clicking on
a bond, element, or arrow, it automatically adjusts their position to be
realistic. But also, fixed an issue where this would unintentionally cause
you to create additional elements. Inadvertently this showed me that you
could also create an arrow by dragging a bond where an arrow should go.
Fixed that too.

And finally, subscript has been added to the element name.
Real-time editing of a JTextField is very, very hard because there are no
provisions for it in the normal libraries, so that is out of the question,
but you can still save a number inside the element name and have it come
back out as subscript.

Signed-off-by: Ian Graham <ianrgraham@gmail.com>